+++ Cargo.toml
@@ -10,6 +10,7 @@ digest = "0.7.2"
 easter = { version = "0.0.5", path = "../esprit/crates/easter" }
 esprit = { version = "0.0.5", path = "../esprit" }
 estree-detect-requires = { path = "crates/estree-detect-requires" }
+node-core-shims = { path = "crates/node-core-shims" }
 node-resolve = "2.0.0"
 serde_json = "1.0"
 sha-1 = "0.7.0"
