+++ crates/node-core-shims/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "node-core-shims"
+version = "0.1.0"
+authors = ["Renée Kooi <renee@kooi.me>"]
+build = "build.rs"
+
+[dependencies]
