+++ crates/node-core-shims/build.rs
@@ -1,0 +1,5 @@
+use std::process::Command;
+
+fn main() {
+    Command::new("npm").args(&["install"]).status().unwrap();
+}
