+++ crates/node-core-shims/package.json
@@ -1,0 +1,23 @@
+{
+  "dependencies": {
+    "assert": "^1.4.1",
+    "browserify-zlib": "^0.2.0",
+    "buffer": "^5.1.0",
+    "crypto-browserify": "^3.12.0",
+    "events": "^2.0.0",
+    "https-browserify": "^1.0.0",
+    "os-browserify": "^0.3.0",
+    "path-browserify": "0.0.0",
+    "process": "^0.11.10",
+    "querystring-es3": "^0.2.1",
+    "readable-stream": "^2.3.6",
+    "stream-browserify": "^2.0.1",
+    "stream-http": "^2.8.1",
+    "string_decoder": "^1.1.1",
+    "timers-browserify": "^2.0.10",
+    "tty-browserify": "0.0.1",
+    "url": "^0.11.0",
+    "util": "^0.10.3",
+    "vm-browserify": "^1.0.1"
+  }
+}
